Add a personal bookkeeping module called mone
The bookkeeper is intended to be used for double-entry bookkeeping. So
money is like a continuum which neither can be created nor
destroyed. It can only change it's owner or like in the case here: an
account.

The bookkeeper can be reached through a web app defined in the
mone.www module.

Signed-off-by: Joe Pearson <pearjo@protonmail.com>